+++ include/boost/decimal/decimal32_t.hpp
@@ -715,7 +715,7 @@ constexpr decimal32_t::decimal32_t(T1 coeff, T2 exp, bool sign) noexcept // NOLI
             // Handle the case of sub-normals that don't need further rounding
             bits_ = sign ? detail::d32_sign_mask : UINT32_C(0); // Reset the sign bit
             const auto zeros {detail::remove_trailing_zeros(reduced_coeff)};
-            biased_exp += zeros.number_of_removed_zeros;
+            biased_exp += static_cast<int>(zeros.number_of_removed_zeros);
             reduced_coeff = zeros.trimmed_number;
             if (biased_exp > 0)
             {
